Fix compile warnings reported by gerrit
The gerrit configuration was incorrectly configured with
respect to warning detection causing the warning score to
be reported incorrectly. This patch address those issues.

Reviewed-on: http://review.couchbase.org/71807
Tested-by: buildbot <build@couchbase.com>

+++ daemon/buckets.h
@@ -221,7 +221,7 @@ extern std::array<Bucket, COUCHBASE_MAX_NUM_BUCKETS + 1> all_buckets;
  */
 void bucketsForEach(std::function<bool(Bucket&, void*)> fn, void *arg);
 
-cJSON *get_bucket_details(int idx);
+cJSON *get_bucket_details(size_t idx);
 
 /**
  * Is the connected bucket currently dying?
