+++ daemon/mcbp_executors.cc
@@ -2092,7 +2092,7 @@ static void get_cmd_timer_executor(McbpConnection* c, void* packet) {
         mcbp_write_and_free(c, &c->getDynamicBuffer());
     } else if (cookie_is_admin(c->getCookie())) {
         bool found = false;
-        for (int ii = 1; ii < all_buckets.size() && !found; ++ii) {
+        for (size_t ii = 1; ii < all_buckets.size() && !found; ++ii) {
             // Need the lock to get the bucket state and name
             cb_mutex_enter(&all_buckets[ii].mutex);
             if ((all_buckets[ii].state == BucketState::Ready) &&
